update tutorial 7 and demo 2
